Revert "Closes #9, Estrategia de login"
