+++ bower.json
@@ -26,7 +26,6 @@
     "c3": "~0.4.10",
     "jquery": "~2.1.4",
     "underscore": "~1.8.3",
-    "socket.io": "~1.3.7",
-    "bootstrap-social": "~4.11.0"
+    "socket.io": "~1.3.7"
   }
 }
