--- config/config-dev.js
@@ -1,4 +1,0 @@
-module.exports = {
-	domain: 'http://localhost:3000',
-	company: 'ondevio'
-}
