--- config/config-pro.js
@@ -1,4 +1,0 @@
-module.exports = {
-	domain: 'https://tricket.ondevio.com',
-	company: 'ondevio'
-}
