+++ package.json
@@ -31,7 +31,6 @@
     "cookie-parser": "~1.3.5",
     "debug": "~2.2.0",
     "express": "~4.13.1",
-    "express-session": "^1.12.1",
     "express.io": "^1.1.13",
     "jade": "~1.11.0",
     "lodash": "^3.10.1",
@@ -39,9 +38,6 @@
     "mongoose": "^4.3.1",
     "morgan": "~1.6.1",
     "node-tail": "0.0.4",
-    "passport": "^0.3.2",
-    "passport-google-oauth": "^0.2.0",
-    "passport-local": "^1.0.0",
     "react": "^0.14.0",
     "react-dom": "^0.14.0",
     "serve-favicon": "~2.3.0",
