+++ routes/index.js
@@ -1,6 +1,14 @@
 var express = require('express');
 var router = express.Router();
 
+router.get('/login', function(req, res, next) {
+  res.render('dashboard/login.html');
+});
+
+router.get('/home', function(req, res, next) {
+  res.render('dashboard/tickets.html');
+});
+
 router.get('/tickets', function(req, res, next) {
   res.render('dashboard/tickets.html');
 });
