--- routes/public.js
@@ -1,50 +1,0 @@
-var express = require('express');
-var router = express.Router();
-var passport = require('passport');
-
-router.get('/', function(req, res, next) {
-  res.redirect('/login');
-});
-
-router.get('/login', function(req, res, next) {
-  res.render('dashboard/login.html');
-});
-
-router.post('/login',
-  passport.authenticate('local', { 
-  	successRedirect: '/app/tickets',
-  	failureRedirect: '/login',
-    failureFlash: true 
-	})
-);
-
-router.get('/logout', function(req, res){
-  req.logout();
-  res.redirect('/login');
-});
-
-// GET /auth/google
-//   Use passport.authenticate() as route middleware to authenticate the
-//   request.  The first step in Google authentication will involve
-//   redirecting the user to google.com.  After authorization, Google
-//   will redirect the user back to this application at /auth/google/callback
-router.get('/auth/google',
-  passport.authenticate('google', { scope: ['https://www.googleapis.com/auth/plus.login', 'https://www.googleapis.com/auth/userinfo.email'] }),
-  function(req, res){
-    // The request will be redirected to Google for authentication, so this
-    // function will not be called.
-  });
-
-// GET /auth/google/callb
-//   Use passport.authenticate() as route middleware to authenticate the
-//   request.  If authentication fails, the user will be redirected back to the
-//   login page.  Otherwise, the primary route function function will be called,
-//   which, in this example, will redirect the user to the home page.
-router.get('/auth/google/callback', 
-  passport.authenticate('google', { failureRedirect: '/login' }),
-  function(req, res) {
-    res.redirect('/app/tickets');
-  });
-
-
-module.exports = router;
