+++ template/dashboard/tickets.html
@@ -49,9 +49,8 @@
       <div class="row">
         <div class="col-sm-3 col-md-2 sidebar">
           <ul class="nav nav-sidebar">
-            <li class="active"><a href="/app/tickets">Tickets <span class="sr-only">(current)</span></a></li>
-            <li><a href="/app/clients">Clients</a></li>
-            <li><a href="/logout">Logout</a></li>
+            <li class="active"><a href="/">Tickets <span class="sr-only">(current)</span></a></li>
+            <li><a href="/clients">Clients</a></li>
           </ul>
         </div>
         <div class="col-sm-9 col-sm-offset-3 col-md-10 col-md-offset-2 main">
